+++ backend/build.gradle
@@ -61,6 +61,8 @@ dependencies {
 	implementation 'it.ozimov:embedded-redis:0.7.1'
 	// s3
 	implementation 'org.springframework.cloud:spring-cloud-starter-aws:2.2.6.RELEASE'
+	// mongo
+	implementation 'org.springframework.boot:spring-boot-starter-data-mongodb'
 }
 
 tasks.named('test') {
