+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/chat/controller/ChatController.java
@@ -1,29 +1,21 @@
 package com.NBE_4_5_2.Team5.domain.chat.controller;
 
 import com.NBE_4_5_2.Team5.domain.chat.entity.ChatMessage;
-import com.NBE_4_5_2.Team5.domain.chat.repository.ChatRoomRepository;
+import com.NBE_4_5_2.Team5.domain.chat.service.ChatRoomService;
 import com.NBE_4_5_2.Team5.domain.chat.service.ChatService;
 import com.NBE_4_5_2.Team5.domain.user.service.AuthTokenService;
-import com.NBE_4_5_2.Team5.domain.user.service.UserService;
-import com.NBE_4_5_2.Team5.global.standard.util.Ut;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.stereotype.Controller;
-
-import java.util.HashMap;
-import java.util.Map;
 
 @Slf4j
 @RequiredArgsConstructor
 @Controller
 public class ChatController {
 
-    private final ChatRoomRepository chatRoomRepository;
+    private final ChatRoomService chatRoomService;
     private final ChatService chatService;
     private final AuthTokenService authTokenService;
 
@@ -38,7 +30,7 @@ public class ChatController {
         // 로그인 회원 정보로 대화명 설정
         message.setSender(nickname);
         // 채팅방 인원수 세팅
-        message.setUserCount(chatRoomRepository.getUserCount(message.getRoomId()));
+        message.setUserCount(chatRoomService.getUserCount(message.getRoomId()));
         // Websocket에 발행된 메시지를 redis로 발행(publish)
         chatService.sendChatMessage(message);
     }
