+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/chat/service/ChatService.java
@@ -1,7 +1,6 @@
 package com.NBE_4_5_2.Team5.domain.chat.service;
 
 import com.NBE_4_5_2.Team5.domain.chat.entity.ChatMessage;
-import com.NBE_4_5_2.Team5.domain.chat.repository.ChatRoomRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.listener.ChannelTopic;
@@ -15,7 +14,7 @@ public class ChatService {
 
     private final ChannelTopic channelTopic;
     private final RedisTemplate redisTemplate;
-    private final ChatRoomRepository chatRoomRepository;
+    private final ChatRoomService chatRoomService;
 
     /**
      * destination정보에서 roomId 추출
@@ -32,7 +31,7 @@ public class ChatService {
      * 채팅방에 메시지 발송
      */
     public void sendChatMessage(ChatMessage chatMessage) {
-        chatMessage.setUserCount(chatRoomRepository.getUserCount(chatMessage.getRoomId()));
+        chatMessage.setUserCount(chatRoomService.getUserCount(chatMessage.getRoomId()));
         if (ChatMessage.MessageType.ENTER.equals(chatMessage.getType())) {
             chatMessage.setMessage(chatMessage.getSender() + "님이 방에 입장했습니다.");
             chatMessage.setSender("[알림]");
